+++ lib/src/customer/sep_12.dart
@@ -61,7 +61,10 @@ class Sep12 {
   /// [memo] and [transactionId] are also optional parameters. See the
   /// Specification on SEP-12 definition.
   Future<AddCustomerResponse> add(Map<String, String> sep9Info,
-      {Map<String, Uint8List>? sep9Files, String? memo, String? type, String? transactionId}) async {
+      {Map<String, Uint8List>? sep9Files,
+      String? memo,
+      String? type,
+      String? transactionId}) async {
     flutter_sdk.PutCustomerInfoRequest request =
         flutter_sdk.PutCustomerInfoRequest();
     request.jwt = token.jwt;
@@ -81,7 +84,10 @@ class Sep12 {
   /// optionally be passed. [memo] and [transactionId] are also optional parameters. See the
   /// Specification on SEP-12 definition.
   Future<AddCustomerResponse> update(Map<String, String> sep9Info, String id,
-      {Map<String, Uint8List>? sep9Files, String? memo, String? type, String? transactionId}) async {
+      {Map<String, Uint8List>? sep9Files,
+      String? memo,
+      String? type,
+      String? transactionId}) async {
     flutter_sdk.PutCustomerInfoRequest request =
         flutter_sdk.PutCustomerInfoRequest();
     request.jwt = token.jwt;
@@ -130,31 +136,36 @@ class AddCustomerResponse {
 
 class GetCustomerResponse {
   String? id;
-  Sep12Status? sep12Status;
-  Map<String, Field?>? fields;
-  Map<String, ProvidedField?>? providedFields;
+  Sep12Status sep12Status;
+  Map<String, Field>? fields;
+  Map<String, ProvidedField>? providedFields;
   String? message;
+
+  GetCustomerResponse(
+      {required this.sep12Status,
+      this.id,
+      this.fields,
+      this.providedFields,
+      this.message});
 
   static GetCustomerResponse from(
       flutter_sdk.GetCustomerInfoResponse infoResponse) {
-    GetCustomerResponse result = GetCustomerResponse();
+    GetCustomerResponse result =
+        GetCustomerResponse(sep12Status: Sep12Status(infoResponse.status));
     result.id = infoResponse.id;
-    if (infoResponse.status != null) {
-      result.sep12Status = Sep12Status(infoResponse.status!);
-    }
 
     if (infoResponse.fields != null) {
-      result.fields = <String, Field?>{};
+      result.fields = <String, Field>{};
       for (String key in infoResponse.fields!.keys) {
-        result.fields![key] = Field.from(infoResponse.fields![key]);
+        result.fields![key] = Field.from(infoResponse.fields![key]!);
       }
     }
 
     if (infoResponse.providedFields != null) {
-      result.providedFields = <String, ProvidedField?>{};
+      result.providedFields = <String, ProvidedField>{};
       for (String key in infoResponse.providedFields!.keys) {
         result.providedFields![key] =
-            ProvidedField.from(infoResponse.providedFields![key]);
+            ProvidedField.from(infoResponse.providedFields![key]!);
       }
     }
     result.message = infoResponse.message;
@@ -163,19 +174,15 @@ class GetCustomerResponse {
 }
 
 class Field {
-  FieldType? type;
+  FieldType type;
   String? description;
   List<String>? choices;
   bool? optional;
 
-  Field({this.type, this.description, this.choices, this.optional});
+  Field({required this.type, this.description, this.choices, this.optional});
 
-  static Field? from(flutter_sdk.GetCustomerInfoField? infoField) {
-    if (infoField == null) {
-      return null;
-    }
-    Field result = Field();
-    result.type = FieldType.fromString(infoField.type);
+  static Field from(flutter_sdk.GetCustomerInfoField infoField) {
+    Field result = Field(type: FieldType.fromString(infoField.type));
     result.description = infoField.description;
     result.choices = infoField.choices;
     result.optional = infoField.optional;
@@ -184,7 +191,7 @@ class Field {
 }
 
 class ProvidedField {
-  FieldType? type;
+  FieldType type;
   String? description;
   List<String>? choices;
   bool? optional;
@@ -192,20 +199,17 @@ class ProvidedField {
   String? error;
 
   ProvidedField(
-      {this.type,
+      {required this.type,
       this.description,
       this.choices,
       this.optional,
       this.status,
       this.error});
 
-  static ProvidedField? from(
-      flutter_sdk.GetCustomerInfoProvidedField? infoProvidedField) {
-    if (infoProvidedField == null) {
-      return null;
-    }
-    ProvidedField result = ProvidedField();
-    result.type = FieldType.fromString(infoProvidedField.type);
+  static ProvidedField from(
+      flutter_sdk.GetCustomerInfoProvidedField infoProvidedField) {
+
+    ProvidedField result = ProvidedField(type: FieldType.fromString(infoProvidedField.type));
     result.description = infoProvidedField.description;
     result.choices = infoProvidedField.choices;
     result.optional = infoProvidedField.optional;
@@ -244,10 +248,8 @@ class FieldType {
   @override
   int get hashCode => Object.hash(value, value);
 
-  static FieldType? fromString(String? typeString) {
-    if (string.value == typeString) {
-      return string;
-    }
+  static FieldType fromString(String typeString) {
+
     if (binary.value == typeString) {
       return binary;
     }
@@ -257,7 +259,8 @@ class FieldType {
     if (date.value == typeString) {
       return date;
     }
-    return null;
+
+    return string;
   }
 }